+++ src/lib/testing_api_cmd_fakebank_transfer.c
@@ -195,7 +195,7 @@ static void
 add_incoming_cb (void *cls,
                  unsigned int http_status,
 		 enum TALER_ErrorCode ec,
-                 uint64_t serial_id,
+                 long long unsigned int serial_id,
                  struct GNUNET_TIME_Absolute timestamp,
                  const json_t *full_response)
 {
